+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+import { formatDate } from "./format-date";
+
+export { formatDate };
